client/requirements.txt
requests>=2.25.0
OPi.GPIO>=0.5.0